requirements.txt
streamlit
openai
requests_html
html2text
llama-index
watchdog
pypdf
beautifulsup4